Implemented right jump order for conditionals

+++ tests/conditionals_gt_false.cc
@@ -6,7 +6,7 @@
 using namespace std;
 
 int main(int argc, char** argv) {
-    int expection = 0.0;
+    double expection = 0.0;
 
     setenv("IMPALA_FILE","impala_files/conditional_gt_false.impala", 1);
