+++ tests/conditionals_gte_true_1.cc
@@ -6,7 +6,7 @@
 using namespace std;
 
 int main(int argc, char** argv) {
-    int expection = 1.0;
+    double expection = 1.0;
 
     setenv("IMPALA_FILE","impala_files/conditional_gte_true_1.impala", 1);
